2025-12-22 PSXMAP v1

+++ PSXMAP.h
@@ -8,7 +8,7 @@
  struct pos{
     double latitude;
     double longitude;
-    uint64_t heading;
+    double heading;
 } Pos;
 
 typedef struct DLNode {
